# Optimize Stream Intersection Calculation