## Overview

Previously we would calculate the intersection of all streams within an area of interest and sum their lengths. For large areas of interest, especially with the high resolution streams, this can be an expensive operation because of the number of streams they cover.

To optimize this, for each intersecting stream, we check if it is entirely covered by the area of interest, using `ST_CoveredBy`. This is a much faster operation than calculating the intersection via `ST_Intersection`. If it is completely covered, we use the full stream. If it is not completely covered, then we calculate the intersection.

The speedup offered by this depends on the size of the AoI and the number of streams within it. As a benchmark I used the Schuylkill HUC-8. Before this change the stream calculation query took about 40s, now it takes less than 20s.

For more details, see: https://postgis.net/2014/03/14/tip_intersection_faster/

Connects #3424 
Connects #3465 

### Demo

```bash
time PGPASSWORD=mmw psql -h 33.33.34.30 -d mmw -U mmw < baseline.sql
      lengthkm      | stream_order |      slopesum
--------------------+--------------+---------------------
  2678.525594519314 |            1 |   68.84731109209834
  925.9671925679618 |            2 |  11.177798795304627
  522.3008888056895 |            3 |   3.141855818161298
  310.0939573096803 |            4 |  0.9727859089283545
  154.9829529552899 |            5 |  0.2500951203820861
  110.0573434427563 |            6 |  0.0837694750544295
 54.803254152631155 |            7 | 0.03016251957984683
 42.199541422473274 |              |
(8 rows)


________________________________________________________
Executed in   43.72 secs      fish           external
   usr time   40.65 millis  752.00 micros   39.90 millis
   sys time    8.24 millis  226.00 micros    8.01 millis
```

```bash
time PGPASSWORD=mmw psql -h 33.33.34.30 -d mmw -U mmw < smarter-intersection.sql
      lengthkm      | stream_order |      slopesum
--------------------+--------------+---------------------
  2678.525594519315 |            1 |   68.84731109209844
  925.9671925679615 |            2 |  11.177798795304625
  522.3008888056894 |            3 |  3.1418558181613006
 310.09395730968004 |            4 |  0.9727859089283541
 154.98295295528985 |            5 | 0.25009512038208614
 110.05734344275628 |            6 | 0.08376947505442944
  54.80325415263115 |            7 | 0.03016251957984683
 42.199541422473274 |              |
(8 rows)


________________________________________________________
Executed in   19.33 secs      fish           external
   usr time   43.57 millis  667.00 micros   42.90 millis
   sys time   10.08 millis    0.00 micros   10.08 millis
```

The slight variation in floating point details is noise and can be ignored.

### Notes

In addition to the smarter intersection technique above, I also tried a variation with lesser deserialization, where the `ST_GeomFromGeoJSON` was captured within a single helper table:

```sql
WITH aoi AS (
        SELECT ST_SetSRID(ST_GeomFromGeoJSON(%s), 4326) AS shape
    ),
    stream_intersection AS (
        SELECT ST_Length(ST_Transform(
                    CASE
                        WHEN ST_CoveredBy(geom, aoi.shape)
                        THEN geom
                        ELSE ST_Intersection(geom, aoi.shape)
                    END,
                    5070)) AS lengthm,
                stream_order,
                slope
        FROM nhdflowlinehr, aoi
        WHERE ST_Intersects(geom, aoi.shape))

SELECT SUM(lengthm) / 1000 AS lengthkm,
        stream_order,
        SUM(lengthm * NULLIF(slope, -9998.0)) / 1000 AS slopesum
FROM stream_intersection
GROUP BY stream_order;
```

However, this did not result in any improvement in execution time:

```bash
time PGPASSWORD=mmw psql -h 33.33.34.30 -d mmw -U mmw < less-serialization.sql
      lengthkm      | stream_order |       slopesum
--------------------+--------------+----------------------
 2678.5255945193117 |            1 |    68.84731109209828
  925.9671925679612 |            2 |   11.177798795304627
  522.3008888056893 |            3 |      3.1418558181613
 310.09395730968015 |            4 |   0.9727859089283548
 154.98295295528982 |            5 |   0.2500951203820862
 110.05734344275629 |            6 |   0.0837694750544295
 54.803254152631155 |            7 | 0.030162519579846822
 42.199541422473274 |              |
(8 rows)


________________________________________________________
Executed in   19.42 secs      fish           external
   usr time   43.92 millis    0.00 micros   43.92 millis
   sys time    5.54 millis  820.00 micros    4.72 millis
```

and could potentially use more RAM as every row of the streams table is joined with a long AoI geometry, so decided against it.

## Testing Instructions

To check validity of these changes:

* Check out this branch and run `./scripts/debugcelery.sh`
* Go to http://localhost:8000/ and run a very large shape, such as a HUC-8
  - [x] Ensure it completes

To check the speed-up:

* Download these files: [baseline.sql.txt](https://github.com/WikiWatershed/model-my-watershed/files/7975635/baseline.sql.txt) and [smarter-intersection.sql.txt](https://github.com/WikiWatershed/model-my-watershed/files/7975636/smarter-intersection.sql.txt)
* Rename them to not have the `.txt` extension (which was added to allow for uploading to GitHub PRs)
* Run the queries via `psql` and `time` them:
    ```bash
    time PGPASSWORD=mmw psql -h 33.33.34.30 -d mmw -U mmw < baseline.sql
    ```
    ```bash
    time PGPASSWORD=mmw psql -h 33.33.34.30 -d mmw -U mmw < smarter-intersection.sql
    ```
  - [x] Ensure the output numbers are similar enough (up to 5 digits of decimal place)
  - [x] Ensure the `smarter-intersection` query takes less time